change 'Test' in 'Makefile' and dependencies

+++ home_tasks/compareSetsWithSort/lib/parsing_from_console.h
@@ -8,7 +8,7 @@
 int_duple_dynamic_array_t get_array_from_console();
 int parse_input_array(char *str, int_duple_dynamic_array_t *array);
 int count_items(char *str);
-int get_number_from_str(char *ch);
+long long get_number_from_str(char *ch);
 int handle_char(char *ch, int_duple_dynamic_array_t *array, int index, int *opening_counter);
 
 #endif //PARSING
